Reformatting db calls to props to remove multiple from the admin dashboard WIP
