+++ itoapp/src/components/taskList/TaskList.js
@@ -1,0 +1,44 @@
+import React from "react";
+import Tasks from "../tasks/Tasks.component";
+import AddTaskLink from "../addTaskLink/AddTaskLink.component";
+import "./TaskList.scss";
+
+export default function TasksList(props) {
+    console.log(props.tasks);
+  return (
+    <div className="tasks">
+      <ol className="uk-list uk-list-striped uk-list-medium">
+        <AddTaskLink />
+        <li className="list-header">
+          <div className="uk-grid">
+            <div className="uk-width-1-3">
+              <p>
+                <strong>Task Name</strong>
+              </p>
+            </div>
+            <div className="uk-width-2-3">
+              <div className="uk-grid">
+                <div className="uk-width-1-3">
+                  <p>
+                    <strong>Assigned to</strong>
+                  </p>
+                </div>
+                <div className="uk-width-1-3">
+                  <p>
+                    <strong>Assigned Date</strong>
+                  </p>
+                </div>
+                <div className="uk-with-1-3">
+                  <p>
+                    <strong>Task Status</strong>
+                  </p>
+                </div>
+              </div>
+            </div>
+          </div>
+        </li>
+        <Tasks tasks={props.tasks} />
+      </ol>
+    </div>
+  );
+}
