+++ itoapp/src/components/tasks/Tasks.component.js
@@ -1,31 +1,8 @@
-import React, { useState, useEffect, useContext } from "react";
-import db from "../../config/firebaseConfig";
+import React from "react";
 import Task from "../task/Task.component";
-import { AuthContext } from "../auth/Auth";
 
-export default function Tasks() {
-  const { userData } = useContext(AuthContext);
-  const [tasks, setTasks] = useState([]);
-
-  useEffect(() => {
-    // Get the tasks
-    let tasks = db
-        .collection("tasks")
-        .where("authid", "==", userData.authid)
-        .orderBy("createdon");
-        
-        let unsubscribe = tasks.onSnapshot(snapshot => {
-        setTasks(
-          snapshot.docs.map(doc => {
-            let task = doc.data();
-            task.id = doc.id;
-            return task;
-          })
-        );
-      });
-
-      return () => unsubscribe();
-  }, [userData.authid]);
+export default function Tasks(props) {
+  const tasks = props.tasks;
 
   return tasks.map((task, index) => {
     if(!task.completed) {
